+++ Tests/SnippetTests.swift
@@ -30,10 +30,11 @@ final class SnippetTests: XCTestCase {
     
     func testSimpleSnippet() {
         
-        let snippet = Snippet(name: "", format: "<h1><<<CURSOR>>></h1>")
+        let snippet = Snippet(name: "", format: "<h1><<<SELECTION>>><<<CURSOR>>></h1>")
+        let (string, selections) = snippet.insertion(selectedString: "abc")
         
-        XCTAssertEqual(snippet.string, "<h1></h1>")
-        XCTAssertEqual(snippet.selections, [NSRange(location: 4, length: 0)])
+        XCTAssertEqual(string, "<h1>abc</h1>")
+        XCTAssertEqual(selections, [NSRange(location: 7, length: 0)])
     }
     
     
@@ -46,6 +47,7 @@ final class SnippetTests: XCTestCase {
             </ul>
             """
         let snippet = Snippet(name: "", format: format)
+        let (string, selections) = snippet.insertion(selectedString: "")
         
         let expectedString = """
             <ul>
@@ -53,11 +55,11 @@ final class SnippetTests: XCTestCase {
                 <li></li>
             </ul>
             """
-        XCTAssertEqual(snippet.string, expectedString)
-        XCTAssertEqual(snippet.selections, [NSRange(location: 13, length: 0),
-                                            NSRange(location: 27, length: 0)])
+        XCTAssertEqual(string, expectedString)
+        XCTAssertEqual(selections, [NSRange(location: 13, length: 0),
+                                    NSRange(location: 27, length: 0)])
         
-        let indentedSnippet = snippet.indented(with: "    ")
+        let (indentedString, indentedSelections) = snippet.insertion(selectedString: "", indent: "    ")
         
         let expectedIndentString = """
             <ul>
@@ -65,8 +67,35 @@ final class SnippetTests: XCTestCase {
                     <li></li>
                 </ul>
             """
-        XCTAssertEqual(indentedSnippet.string, expectedIndentString)
-        XCTAssertEqual(indentedSnippet.selections, [NSRange(location: 17, length: 0),
-                                                    NSRange(location: 35, length: 0)])
+        XCTAssertEqual(indentedString, expectedIndentString)
+        XCTAssertEqual(indentedSelections, [NSRange(location: 17, length: 0),
+                                            NSRange(location: 35, length: 0)])
+    }
+    
+    
+    func testMultipleInsertions() {
+        
+        let string = """
+                aaa
+            
+            bbcc
+            """
+        let snippet = Snippet(name: "", format: "<li><<<SELECTION>>><<<CURSOR>>></li>")
+        let (strings, selections) = snippet.insertions(for: string, ranges: [
+            NSRange(location: 4, length: 3),
+            NSRange(location: 8, length: 0),
+            NSRange(location: 9, length: 2),
+        ])
+        
+        let expectedStrings = [
+            "<li>aaa</li>",
+            "<li></li>",
+            "<li>bb</li>",
+        ]
+        let expectedSelections = [NSRange(location: 11, length: 0),
+                                  NSRange(location: 21, length: 0),
+                                  NSRange(location: 33, length: 0)]
+        XCTAssertEqual(strings, expectedStrings)
+        XCTAssertEqual(selections, expectedSelections)
     }
 }